Fixes chat.post_message with attachments
The chat.postMessage API expects things to be passed to it as params,
not data. Without this, attachments won't be posted.

This also removes any keys that are None so they aren't sent to Slack.

It also makes the text argument optional, since it the API specifies you
text is only required if you aren't specifying attachments. You'll get a
slack.Error back from the API if you don't specify either.